+++ tests/tests2/117_builtins.expect
@@ -1,0 +1,4 @@
+BOUNDS OFF:
+ 1:1 2:1 3:1 4:1 5:1 6:1 7:1 8:1 9:1 10:1 11:1 12:1 13:1
+BOUNDS ON:
+ 1:1 2:1 3:1 4:1 5:1 6:1 7:1 8:1 9:1 10:1 11:1 12:1 13:1
