+++ tests/tests2/118_switch.c
